refactor(ir): Split FlattenTileNdTo2D into phases
